requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
cloudinary==1.30.0
dj-database-url==1.0.0
Django==4.1.2
django-autoslug==1.9.8
django-ckeditor==6.5.1
django-cloudinary-storage==0.3.0
django-debug-toolbar==3.7.0
django-jazzmin==2.5.0
django-js-asset==2.0.0
django-social-share==2.3.0
gunicorn==20.1.0
idna==3.4
Naked==0.1.31
Pillow==9.2.0
psycopg2-binary==2.9.4
python-decouple==3.6
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
sweetify==2.3.1
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
whitenoise==6.2.0
